Move registration helper functions to _registration_utils
