+++ fractal_tasks_core/tasks/calculate_registration_image_based.py
@@ -17,7 +17,6 @@ import logging
 import anndata as ad
 import dask.array as da
 import numpy as np
-import pandas as pd
 import zarr
 from pydantic.decorator import validate_arguments
 from skimage.registration import phase_cross_correlation
@@ -34,6 +33,12 @@ from fractal_tasks_core.roi import (
 )
 from fractal_tasks_core.roi import load_region
 from fractal_tasks_core.tables import write_table
+from fractal_tasks_core.tasks._registration_utils import (
+    calculate_physical_shifts,
+)
+from fractal_tasks_core.tasks._registration_utils import (
+    get_ROI_table_with_translation,
+)
 from fractal_tasks_core.tasks.io_models import InitArgsRegistration
 
 logger = logging.getLogger(__name__)
@@ -89,8 +94,6 @@ def calculate_registration_image_based(
 
     # Get channel_index via wavelength_id.
     # Intially only allow registration of the same wavelength
-    print(init_args.reference_zarr_url)
-    print(zarr_url)
     channel_ref: OmeroChannel = get_channel_from_image_zarr(
         image_zarr_path=init_args.reference_zarr_url,
         wavelength_id=wavelength_id,
@@ -251,75 +254,6 @@ def calculate_registration_image_based(
         overwrite=True,
         table_attrs=ref_table_attrs,
     )
-
-
-def calculate_physical_shifts(
-    shifts: np.array,
-    level: int,
-    coarsening_xy: int,
-    full_res_pxl_sizes_zyx: list[float],
-) -> list[float]:
-    """
-    Calculates shifts in physical units based on pixel shifts
-
-    Args:
-        shifts: array of shifts, zyx or yx
-        level: resolution level
-        coarsening_xy: coarsening factor between levels
-        full_res_pxl_sizes_zyx: pixel sizes in physical units as zyx
-
-    Returns:
-        shifts_physical: shifts in physical units as zyx
-    """
-
-    curr_pixel_size = np.array(full_res_pxl_sizes_zyx) * coarsening_xy**level
-    if len(shifts) == 3:
-        shifts_physical = shifts * curr_pixel_size
-    elif len(shifts) == 2:
-        shifts_physical = [
-            0,
-            shifts[0] * curr_pixel_size[1],
-            shifts[1] * curr_pixel_size[2],
-        ]
-    else:
-        raise ValueError(
-            f"Wrong input for calculate_physical_shifts ({shifts=})"
-        )
-    return shifts_physical
-
-
-def get_ROI_table_with_translation(
-    ROI_table: ad.AnnData,
-    new_shifts: dict[str, list[float]],
-) -> ad.AnnData:
-    """
-    Adds translation columns to a ROI table
-
-    Args:
-        ROI_table: Fractal ROI table
-        new_shifts: zyx list of shifts
-
-    Returns:
-        Fractal ROI table with 3 additional columns for calculated translations
-    """
-
-    shift_table = pd.DataFrame(new_shifts).T
-    shift_table.columns = ["translation_z", "translation_y", "translation_x"]
-    shift_table = shift_table.rename_axis("FieldIndex")
-    new_roi_table = ROI_table.to_df().merge(
-        shift_table, left_index=True, right_index=True
-    )
-    if len(new_roi_table) != len(ROI_table):
-        raise ValueError(
-            "New ROI table with registration info has a "
-            f"different length ({len(new_roi_table)=}) "
-            f"from the original ROI table ({len(ROI_table)=})"
-        )
-
-    adata = ad.AnnData(X=new_roi_table.astype(np.float32))
-    adata.obs_names = new_roi_table.index
-    adata.var_names = list(map(str, new_roi_table.columns))
-    return adata
 
 
 if __name__ == "__main__":
